requirements.txt
chevron
pydantic
PyYAML
diagrams
oscalic @ git+https://github.com/Compton-NIST/OSCALic.git@develop